+++ corona-ide-core/src/main/java/com/coronaide/core/impl/package-info.java
@@ -1,0 +1,19 @@
+/*******************************************************************************
+ * Copyright (c) Oct 24, 2016 Corona IDE.
+ * All rights reserved. This program and the accompanying materials
+ * are made available under the terms of the Eclipse Public License v1.0
+ * which accompanies this distribution, and is available at
+ * http://www.eclipse.org/legal/epl-v10.html
+ *
+ * Contributors:
+ *    romeara - initial API and implementation and/or initial documentation
+ *******************************************************************************/
+/**
+ * Implementations of representations defined in the core APIs
+ *
+ * @author romeara
+ */
+@ParametersAreNonnullByDefault
+package com.coronaide.core.impl;
+
+import javax.annotation.ParametersAreNonnullByDefault;